Move SelectableAdapter to package
