+++ src/Adapter/DoctrineCollectionAdapter.php
@@ -4,7 +4,7 @@ namespace Pagerfanta\Adapter;
 
 use Pagerfanta\DoctrineCollections\DoctrineCollectionsAdapter;
 
-trigger_deprecation('pagerfanta/pagerfanta', '2.2', 'The "%s" class is deprecated and will be removed in 3.0. Use the "%s" class from the "pagerfanta/doctrine-collections-adapter" package instead.', DoctrineCollectionAdapter::class, DoctrineCollectionsAdapter::class);
+trigger_deprecation('pagerfanta/pagerfanta', '2.4', 'The "%s" class is deprecated and will be removed in 3.0. Use the "%s" class from the "pagerfanta/doctrine-collections-adapter" package instead.', DoctrineCollectionAdapter::class, DoctrineCollectionsAdapter::class);
 
 /**
  * Adapter which calculates pagination from a Doctrine Collection.
